Fix correctness, hardening, and test-quality findings from full-project review
Addresses verified findings from a whole-project review, across four tiers.
The Data/Spacer shared-LP-variable finding is intentionally deferred to the
in-flight post-synthesis extension; the two pseudo-clock findings and the
mutex self-deadlock were verified as false positives and left unchanged.

Correctness:
- validate_hbcn: reject duplicate parallel places (same transition pair)
- structural graph: raise MultipleDefinitions for redefined DataReg/UnsafeReg
  intermediate nodes; reject the reserved `port:` prefix on non-port names
- grammars: reject overlong numeric literals that parse to f64 infinity
- analyse depth table: print real f64 slack instead of truncating to usize
- demo/comment: correct log filename (lp_solver.log -> hbcn.log)

Hardening:
- coin_cbc: extract solution values only when proven optimal (mirror Gurobi);
  coalesce duplicate variables so set_weight/set_obj_coeff sum instead of
  overwrite; guard round_to_sig_digits against NaN/inf; scale strict-`>`
  epsilon to RHS magnitude
- ops: fix reverse `Sub<VariableId> for f64` sign (scalar - var)

Panic hardening:
- vcd: NaN-safe total_cmp when sorting events
- analyse/constrain: handle bellman_ford / find_edge failures gracefully

Tests:
- delete dead+stale sdc_tests.rs; recover its unique coverage (src/dst rails,
  precision, min-only/empty, TCL structure) as real tests against the current API
- make the infeasible integration test and forward-completion test assert their
  intended behaviour instead of nothing
- add regression tests for the new validations and the coalesce/sign fixes

Also folds in pre-existing fmt/lint/CI housekeeping (make fmt/clippy real CI
gates, drop a redundant import, trailing-whitespace).

+++ src/analyse/mod.rs
@@ -118,7 +118,7 @@ pub struct AnalyseArgs {
 /// ```
 pub fn analyse_main(args: AnalyseArgs) -> Result<()> {
     use crate::output_suppression::is_verbose;
-    
+
     let AnalyseArgs {
         input,
         structural,
@@ -154,7 +154,7 @@ pub fn analyse_main(args: AnalyseArgs) -> Result<()> {
             let g = read_file(&input)?;
             let hbcn = crate::hbcn::from_structural_graph(&g, false)
                 .ok_or_else(|| anyhow!("Failed to convert structural graph to StructuralHBCN"))?;
-            
+
             if is_verbose() {
                 eprintln!("Computing cycle time (weighted={})...", weighted);
             }
@@ -164,7 +164,7 @@ pub fn analyse_main(args: AnalyseArgs) -> Result<()> {
             let file_contents = fs::read_to_string(&input)?;
             let hbcn = crate::hbcn::parser::parse_hbcn(&file_contents)?;
             // DelayedPlace implements HasWeight, so we can use it directly
-            
+
             if is_verbose() {
                 eprintln!("Computing cycle time (weighted={})...", weighted);
             }
@@ -337,7 +337,7 @@ pub fn analyse_main(args: AnalyseArgs) -> Result<()> {
                     },
                     s.transition.name(),
                     ttype,
-                    format!("{}", e.slack() as usize),
+                    format!("{}", e.slack()),
                     format!("{}", s.time),
                 ]);
             }
